+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog for PSRemoteOperations
 
+## v0.4.0
+
++ Added initialization script option to RemoteJob definition and execution
++ Added `New-PSRemoteProfileScript` function
++ Added `PSScheduledJob` as a required module
++ Set default name for `-Credential` parameter of `Register-PSRemoteOperationWatcher` to `$env:username`
++ Added help documentation
++ Created Pester tests for the module
+
 ## v0.3.0
 
 + Renamed function nouns to include a PSPrefix